Solved ships moving way too fast. Also removed a few unneeded cast to floats. Ships were moving too fast because the 'Mass' variable is inherited from 'GameplayObject'. I just had to have the ShipModule_Deserialize inherit 'GameplayObject' as well, and copy the 'Mass' read in from the XML files to it.
--HG--
branch : feature/Memory Usage Overhall

+++ Ship_Game/Gameplay/ShipModule.cs
@@ -120,8 +120,8 @@ namespace Ship_Game.Gameplay
         public bool IsTroopBay                  { get { return this.Advanced.IsTroopBay; } }
         public float hangarTimerConstant        { get { return this.Advanced.hangarTimerConstant; } }
         public float thrust                     { get { return this.Advanced.thrust; } }
-        public int WarpThrust                   { get { return this.Advanced.WarpThrust; } }
-        public int TurnThrust                   { get { return this.Advanced.TurnThrust; } }
+        public float WarpThrust                 { get { return this.Advanced.WarpThrust; } }
+        public float TurnThrust                 { get { return this.Advanced.TurnThrust; } }
         public float PowerFlowMax               { get { return this.Advanced.PowerFlowMax; } }
         public float PowerDraw                  { get { return this.Advanced.PowerDraw; } }
         public float PowerDrawAtWarp            { get { return this.Advanced.PowerDrawAtWarp; } }
